Refactors httpbin controller into its own test class
